Move `InsertionPointsView` into the `backend` module
