+++ backend/src/main/kotlin/io/spine/protodata/backend/InsertionPointsView.kt
@@ -1,11 +1,11 @@
 /*
- * Copyright 2023, TeamDev. All rights reserved.
+ * Copyright 2024, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ * https://www.apache.org/licenses/LICENSE-2.0
  *
  * Redistribution and use in source and/or binary forms, with or without
  * modification, must retain the above copyright notice and the following
@@ -24,18 +24,21 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.protodata
+package io.spine.protodata.backend
 
 import io.spine.core.External
 import io.spine.core.Subscribe
+import io.spine.protodata.File
+import io.spine.protodata.InsertedPoints
 import io.spine.protodata.event.InsertionPointPrinted
+import io.spine.protodata.insertedPoint
 import io.spine.protodata.plugin.View
 import io.spine.server.entity.alter
 
 /**
  * A view of insertion points added to a single code file.
  */
-public class InsertionPointsView : View<File, InsertedPoints, InsertedPoints.Builder>() {
+internal class InsertionPointsView : View<File, InsertedPoints, InsertedPoints.Builder>() {
 
     @Subscribe
     internal fun on(@External e: InsertionPointPrinted) = alter {
